mds: consider 'in' mds when hashing inode/dirfrag into rank
The hash_into_rank_bucket function used get_max_mds() to determine
the number of ranks for distributing inodes and dirfrags.  However,
get_max_mds() returns the configured maximum number of MDS ranks,
which can be larger than the number of ranks actually "in" the
cluster.  When the max_mds exceeds get_num_in_mds(), the hash
function can assign items to ranks that have not yet joined, causing
requests targeted at those ranks to stall until the rank appears, or
to fail entirely if the rank never comes up.  Use the minimum of
get_max_mds() and get_num_in_mds() as the effective rank count, so
that hashing only distributes across ranks that are actually
participating in the filesystem.

Fixes: https://tracker.ceph.com/issues/77427
Signed-off-by: Yan, Zheng <zyan@redhat.com>
Signed-off-by: Xiubo Li <xiubo.li@clyso.com>