mds: don't set xlock of snaplock done when early replying rename
When Locker::set_xlocks_done() is called with skip_dentry=true for an
early reply to a rename, it skipped marking the dversion lock as done
(in addition to the dentry lock).  However, path traversal can be
serialized via either dentry locks or snaplocks.  Marking the
snaplock xlock as done during an early reply would allow a concurrent
operation to proceed past the snaplock while the rename is still
holding it, potentially seeing inconsistent directory state.  Skip
the snaplock instead of the dversion lock so that the rename holds
its snaplock xlock until the operation fully completes, while the
dversion lock (which is not used for path locking) can be released
early.

Fixes: https://tracker.ceph.com/issues/77427
Signed-off-by: Yan, Zheng <zyan@redhat.com>
Signed-off-by: Xiubo Li <xiubo.li@clyso.com>