mds: resync straydir's fragstat/rstat when purging stray
After fragmenting, a stray directory's fragstat and rstat counters
can become stale because the fragmentation process may not
immediately propagate accounting updates to every fragment.  When the
stray manager purges a stray dentry, it directly modifies the
fragstat counters (decrementing nsubdirs for directories, nfiles for
files) based on stale values, which can cause the directory
accounting to drift further.  Call resync_accounted_fragstat() and
resync_accounted_rstat() before modifying the fragstat, so that the
counters are brought up to date first and the subsequent adjustments
are applied against accurate baselines.

Fixes: https://tracker.ceph.com/issues/77427
Signed-off-by: Yan, Zheng <zyan@redhat.com>
Signed-off-by: Xiubo Li <xiubo.li@clyso.com>