mds: don't send pending rejoins if cluster is no longer in rejoining state
When rejoin_send_rejoins() is called with target MDS_RANK_NONE (a
periodic retry), it previously set rejoins_pending at the very end of
the function, after potentially sending rejoins to all peers.  If the
cluster transitioned out of the rejoining state between the time the
retry was scheduled and the time it ran, the MDS would still send
rejoin messages to peers that had already moved on, and the
rejoins_pending flag would remain set until messages were sent.
Check mdsmap->is_rejoining() at the start of the retry path, clear
rejoins_pending early, and bail out with a no-op if the cluster is no
longer in the joint rejoining state.  This avoids spurious rejoin
messages and ensures rejoins_pending accurately reflects whether a
retry is still needed.

Fixes: https://tracker.ceph.com/issues/77427
Signed-off-by: Yan, Zheng <zyan@redhat.com>
Signed-off-by: Xiubo Li <xiubo.li@clyso.com>