mds: avoid assertion after resetting sessionmap
When the MDS replays a journal event that references a preallocated
inode from a session, it asserts that the inode returned by the
session's take_ino() matches the expected value.  This assertion
fails if the sessionmap was previously reset with mds_wipe_sessions,
because the session's prealloc_inos set is reconstructed from the
journal and may no longer contain the exact inode the event refers
to.  Relax the assertion so that it also passes when
mds_wipe_sessions is set, allowing the MDS to complete journal replay
after a sessionmap reset without crashing.

Fixes: https://tracker.ceph.com/issues/77427
Signed-off-by: Yan, Zheng <zyan@redhat.com>
Signed-off-by: Xiubo Li <xiubo.li@clyso.com>