mds: ignore killed session in MDSRank::get_session()
When a session is killed, its connection is reset to null.  The
session replacement logic in get_session() dereferenced
get_connection() unconditionally, crashing on a killed session.
Guard the session replacement path with a null check on
get_connection() so that killed sessions are silently ignored.
Also simplify connection and session pointer access to avoid
redundant .get() calls.

Fixes: https://tracker.ceph.com/issues/77427
Signed-off-by: Yan, Zheng <zyan@redhat.com>
Signed-off-by: Xiubo Li <xiubo.li@clyso.com>