mds: don't hold wrlock while requesting remote wrlock
When acquire_locks() encounters a lock that needs a remote wrlock, it
previously left the local wrlock held while waiting for the remote
peer to grant the remote wrlock.  Holding both the local wrlock and
waiting on a network round-trip creates a window where other
operations on the same object are blocked.  If the remote peer is
also waiting on a lock held by this MDS, this can deadlock.  Drop the
local wrlock via wrlock_finish() before sending the remote wrlock
request, and collect any caps that need to be re-issued as a result
of releasing the lock.  This allows other operations to proceed
locally while the remote wrlock request is in flight.

Fixes: https://tracker.ceph.com/issues/77427
Signed-off-by: Yan, Zheng <zyan@redhat.com>
Signed-off-by: Xiubo Li <xiubo.li@clyso.com>