mds: disable balancer by default
Change mds_bal_max type from int to uint and default from -1
(unlimited iterations) to 0 (disabled).  The balancer is designed for
small-to-medium clusters and can interfere with the no-subtreemap
mode where subtree placement is handled by hash-based distribution
and migration.  Set the default to disabled so that operators must
explicitly opt in.

Fixes: https://tracker.ceph.com/issues/77427
Signed-off-by: Yan, Zheng <zyan@redhat.com>
Signed-off-by: Xiubo Li <xiubo.li@clyso.com>