mds: fix Server::handle_client_openc to handle existing non-auth file
When handle_client_openc() finds that the dentry already exists and
the inode is non-auth, forward the request to the authoritative MDS
instead of proceeding locally.  Also, when the openc has been
forwarded at least twice and the path is not yet locked, try a
regular open path first in case the file already exists elsewhere,
to avoid an unnecessary create attempt.

Fixes: https://tracker.ceph.com/issues/77427
Signed-off-by: Yan, Zheng <zyan@redhat.com>
Signed-off-by: Xiubo Li <xiubo.li@clyso.com>