mds: update the export targets before journaling the EImportStart
When the importer mds crashed just after the EImportStart journal
was flushed, the standby mds will replay it later, and when replaying
the EImportStart journal the standby mds will wait the client to
reconnect, but actually the client may not open the session yet.

This fixing will update the export_targets to mdsmap before flushing
EImportStart journal, then in the Client side we can use this info
to reconnect the export target mds.

Signed-off-by: Xiubo Li <xiubli@redhat.com>