mds: transit scatter locks of replicas to MIX lazily
When a directory scatterlock transitions between LOCK and MIX states,
rather than synchronously sending lock messages to every replica,
defer the MIX transition on replicas that do not have a subtree root
for that inode.  These "lazy" replicas are tracked in a lazy_set on
the ScatterLock and are only transitioned when they later need the
MIX state (e.g. when a subtree is created on that replica).  This
avoids unnecessary lock message traffic for dirfrags that are
replicated but rarely accessed on a given MDS.

Implement the lazy scatter with add_lazy_scatter(),
remove_lazy_scatter(), and the LOCK_AC_UNLAZY / LOCK_AC_MIXSYNC lock
actions.  Replace the inode-level num_subtree_roots counter with a
per-auth subtree_auth_map so that we can efficiently check whether
an inode has a subtree root on a particular MDS.  Factor the
export-time lock twiddling into CInode::twiddle_locks() and add
CInode::remove_lazy_scatter() for cleanup during import reversal and
replica removal.

Fixes: https://tracker.ceph.com/issues/77427
Signed-off-by: Yan, Zheng <zyan@redhat.com>
Signed-off-by: Xiubo Li <xiubo.li@clyso.com>