mds: add truncate size handling support for fscrypt
When the fscrypt is enabled and when truncating a smaller size, the
old size and new size will always rounded up to CEPH_FSCRYPT_BLOCK_SIZE,
which is 4K for now. For example if truncating file size from 3KB
to 2KB, the MDS will always get old_size == new_size == 4KB. But
only in this case will the truncate request pass the encrypted last
block, we can check this and if the length is non-zero we can be
sure that it's truncating to smaller size.

Signed-off-by: Xiubo Li <xiubli@redhat.com>