mds: add LOCK_LOCK_XLOCK2 support for fscrypt
For fscrypt file when truncating a smaller size, which is not aligned
to the FSCRYPT BLOCK size, the truncate request will pass the encrypted
last block to MDS, then the MDS will update it when truncating the size.

We need to make sure no client is buffering the last block contents
during truncating size. Set filelock state to XLOCK2 will revoke
the Fcb from all the clients and the clients will flush the pagecaches.

Signed-off-by: Xiubo Li <xiubli@redhat.com>