mds: update backtrace for COWed inodes
When unlinking a file after being snapshot, it will cow a new inode
and then store it, but without backtrace info in it. So when doing
recovery with an alternate metadata pool when recovering the snaps
we couldn't find the proper snaprealms.

We need to store the backtrace for the COWed inodes.

Fixes: https://tracker.ceph.com/issues/56605
Signed-off-by: Xiubo Li <xiubli@redhat.com>