qa: add new mntargs option for fuse
test_client_recovery was also using mntopts to specify additional
options to ceph-fuse. Because the two prior commits unify the behavior
of ceph-fuse and the kernel mount so that the "-o" option is available
for both, that changes breaks this test. Add a special set of args
available only for fuse (there is no equivalent on the kernel).

Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>

 Conflicts:
	qa/tasks/ceph_fuse.py
	qa/tasks/cephfs/fuse_mount.py