qa: install the dependency packages required by xfsprogs-dev repo
We need to install the xfsprogs-dev from source to make the xfs_io
command to support the 'add_enckey'.

Currently we couldn't install 'userspace-rcu-devel', 'libedit-devel'
and 'device-mapper-devel' from rhel distos. So the file encyrpt test
cases will skip the rhel distos for now. And it works well on centos
and ubuntu.

Signed-off-by: Xiubo Li <xiubli@redhat.com>