client: allow unprivileged users to clear suid/sgid
The fallocate path will try to clear the suid/sgid if a unprevileged
user changed the file. And the VFS will set the ATTR_MODE and at the
same time by clearing the suid/sgid mode bits.

There is no Posix item requires that we should clear the suid/sgid
in fallocate code path but this is the default behaviour for most of
the filesystems and the VFS layer. And also the same for the write
code path, which have already support it.

Signed-off-by: Xiubo Li <xiubli@redhat.com>