client: show the correct snapshots size
For snapdir, when the 'client_dirsize_rbytes' is disabled it will
show the total number of the snapshots under it, else it will show
the total size for all the snapshots.

For snapshots, when the 'client_dirsize_rbytes' is disabled it'll
show the total number of the sub-directories and files under it,
else will show the total size of snapshot.

After 'client_dirsize_rbytes' enabled we need to check whether the
mds has enabled the 'snap_rstat' or not, it not it still couldn't
correctly show the snapdir and snapshots' rstat. Then just report
0 instead.

Fixes: https://tracker.ceph.com/issues/61151
Signed-off-by: Xiubo Li <xiubli@redhat.com>