osd: fix the use-after-move bug for the extent array in sparse-read
The 'm' map, which is the extent array in sparse-read protocol, is
possibly modified by the objects_readv_sync() and also it will be
encoded and passed to the clients together with the sparse-read op
reply message.

Continue using the 'm' map is incorrect after being std::move().
Though the contents in the 'm' map are still valid after being
std::move(), if the contents being modified by objects_readv_sync()
the 'm' map will become stale. Then the real data length and the
totoal length in the extent array will mismatch.

Currently this change will only affect the sparse-read OSD op.

Introduced-by: 7c53a69eaee("include: codify the ownership change of the map").
Fixes: https://tracker.ceph.com/issues/62081
Signed-off-by: Xiubo Li <xiubli@redhat.com>