mds: set the correct WRLOCK flag always in wrlock_force()
When the lock is added as a remote wrlock first it will add one inest
lock item to the 'mdr->locks'. But later when the same mdr tries to
added a wrlock it will fail to set the 'WRLOCK' flag. Because the
emplace() will return the existing item only without setting the
flag.

Fixes: https://tracker.ceph.com/issues/65630
Signed-off-by: Xiubo Li <xiubli@redhat.com>